+++ Hengband/Hengband/Hengband.vcxproj.filters
@@ -1457,12 +1457,6 @@
     <ClCompile Include="..\..\src\load\load-util.cpp">
       <Filter>load</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\load\load-v1-5-0.cpp">
-      <Filter>load</Filter>
-    </ClCompile>
-    <ClCompile Include="..\..\src\load\item-loader.cpp">
-      <Filter>load</Filter>
-    </ClCompile>
     <ClCompile Include="..\..\src\load\monster-loader.cpp">
       <Filter>load</Filter>
     </ClCompile>
@@ -1482,9 +1476,6 @@
       <Filter>load</Filter>
     </ClCompile>
     <ClCompile Include="..\..\src\load\extra-loader.cpp">
-      <Filter>load</Filter>
-    </ClCompile>
-    <ClCompile Include="..\..\src\load\load-v1-7-0.cpp">
       <Filter>load</Filter>
     </ClCompile>
     <ClCompile Include="..\..\src\load\player-info-loader.cpp">
@@ -2363,6 +2354,15 @@
     </ClCompile>
     <ClCompile Include="..\..\src\util\rng-xoshiro.cpp">
       <Filter>util</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\load\old\load-v1-5-0.cpp">
+      <Filter>load\old</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\load\old\load-v1-7-0.cpp">
+      <Filter>load\old</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\load\old\item-loader-savefile10.cpp">
+      <Filter>load\old</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -4070,12 +4070,6 @@
     <ClInclude Include="..\..\src\load\load-util.h">
       <Filter>load</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\load\load-v1-5-0.h">
-      <Filter>load</Filter>
-    </ClInclude>
-    <ClInclude Include="..\..\src\load\item-loader.h">
-      <Filter>load</Filter>
-    </ClInclude>
     <ClInclude Include="..\..\src\load\savedata-flag-types.h">
       <Filter>load</Filter>
     </ClInclude>
@@ -4098,9 +4092,6 @@
       <Filter>load</Filter>
     </ClInclude>
     <ClInclude Include="..\..\src\load\extra-loader.h">
-      <Filter>load</Filter>
-    </ClInclude>
-    <ClInclude Include="..\..\src\load\load-v1-7-0.h">
       <Filter>load</Filter>
     </ClInclude>
     <ClInclude Include="..\..\src\load\player-info-loader.h">
@@ -5121,6 +5112,15 @@
     <ClInclude Include="..\..\src\mspell\mspell.h">
       <Filter>mspell</Filter>
     </ClInclude>
+    <ClInclude Include="..\..\src\load\old\load-v1-5-0.h">
+      <Filter>load\old</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\src\load\old\load-v1-7-0.h">
+      <Filter>load\old</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\src\load\old\item-loader-savefile10.h">
+      <Filter>load\old</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <None Include="..\..\src\wall.bmp" />
@@ -5342,6 +5342,9 @@
     </Filter>
     <Filter Include="timed-effect">
       <UniqueIdentifier>{b9e618e1-f780-4f5f-9413-d0270ac53660}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="load\old">
+      <UniqueIdentifier>{46523373-91fb-436b-89c6-14ad74b88f5d}</UniqueIdentifier>
     </Filter>
   </ItemGroup>
   <ItemGroup>
