+++ src/load/load.cpp
@@ -22,12 +22,12 @@
 #include "load/extra-loader.h"
 #include "load/info-loader.h"
 #include "load/inventory-loader.h"
-#include "load/item-loader.h"
 #include "load/load-util.h"
-#include "load/load-v1-5-0.h"
-#include "load/load-v1-7-0.h"
 #include "load/load-zangband.h"
 #include "load/lore-loader.h"
+#include "load/old/item-loader-savefile10.h"
+#include "load/old/load-v1-5-0.h"
+#include "load/old/load-v1-7-0.h"
 #include "load/option-loader.h"
 #include "load/player-info-loader.h"
 #include "load/quest-loader.h"
