+++ src/load/old/item-loader-savefile10.cpp
@@ -1,9 +1,9 @@
-﻿#include "load/item-loader.h"
+﻿#include "load/old/item-loader-savefile10.h"
 #include "artifact/fixed-art-types.h"
 #include "game-option/runtime-arguments.h"
 #include "load/angband-version-comparer.h"
 #include "load/load-util.h"
-#include "load/load-v1-5-0.h"
+#include "load/old/load-v1-5-0.h"
 #include "load/savedata-flag-types.h"
 #include "load/savedata-old-flag-types.h"
 #include "object-enchant/object-ego.h"
