+++ src/load/player-info-loader.cpp
@@ -5,8 +5,8 @@
 #include "load/birth-loader.h"
 #include "load/dummy-loader.h"
 #include "load/load-util.h"
-#include "load/load-v1-7-0.h"
 #include "load/load-zangband.h"
+#include "load/old/load-v1-7-0.h"
 #include "load/player-attack-loader.h"
 #include "load/player-class-specific-data-loader.h"
 #include "load/world-loader.h"
