+++ src/load/store-loader.cpp
@@ -2,8 +2,8 @@
 #include "avatar/avatar.h"
 #include "floor/floor-town.h"
 #include "load/angband-version-comparer.h"
-#include "load/item-loader.h"
 #include "load/load-util.h"
+#include "load/old/item-loader-savefile10.h"
 #include "object/object-stack.h"
 #include "object/object-value.h"
 #include "store/store.h"
